refactor(CommandMapper): Command 가 View 를 의존하지 않도록 변경
