+++ src/test/java/chess/view/CommandMapperTest.java
@@ -1,11 +1,11 @@
-package chess.command;
+package chess.view;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class CommandTest {
+class CommandMapperTest {
 
     @Test
     @DisplayName("명령이 올바른 형식으로 들어오지 않았을 때 예외를 발생시킨다")
@@ -14,7 +14,7 @@ class CommandTest {
         final String input = "s t a r t";
 
         // then
-        assertThatThrownBy(() -> Command.from(input))
+        assertThatThrownBy(() -> CommandMapper.from(input))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("잘못된 명령입니다.");
     }
@@ -26,7 +26,7 @@ class CommandTest {
         final String input = "move a2";
 
         // then
-        assertThatThrownBy(() -> Command.from(input))
+        assertThatThrownBy(() -> CommandMapper.from(input))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("잘못된 명령입니다.");
     }
